requirements.txt
alpaca-py>=0.13.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
ta>=0.11.0
